requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.4
django-braces==1.11.0
django-cors-middleware==1.3.1
djangorestframework==3.6.3
docutils==0.14
gunicorn==19.7.1
idna==2.6
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jmespath==0.9.3
oauthlib==2.0.4
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.10
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
